feat: add preimage recovery for send swaps
[skip ci]
